Implemented a factory to generate a minefield of appropriate dimensions
At the moment this type isn't doing anything except delegating to the constructor of the Minefield type so we _could_ just call that directly. However, we will soon need to do things such as place mines in certain locations.

This code is supposed to be a fair reflection of the process I went through when first presented with this problem, rather than just showing what the final version of the code was.