Removed the dummy test
We no longer need this to prove we can run tests as we have actual tests we can run!

This code is supposed to be a fair reflection of the process I went through when first presented with this problem, rather than just showing what the final version of the code was.