Added more test cases for the Minefield.GetHint method.
This was achieved by using `Theory` for the existing test, and some `InlineData`, rather than just a `Fact`.

This code is supposed to be a fair reflection of the process I went through when first presented with this problem, rather than just showing what the final version of the code was.