Improved performance of the Minefield.IsMined method
Used the fact that `Record` gives us a good implementation of `GetHashCode`. We can put `Squares` into a `HashSet<Square>` so we can quickly look up whether the `Square` is contained, rather than having to enumerate potentially hundreds, or even thousands of `Square`s manually checking each one.

Thanks to Chris and Raees for the steer on this one 🙂

This code is supposed to be a fair reflection of the process I went through when first presented with this problem, rather than just showing what the final version of the code was.