Renamed Square to Cell
We would expect a square to have properties such as `Length`, rather than `Row` and `Column`. `Cell` seems like a better choice of name as the information we are interested in is the intersection of row, and column, _not_ the shape of a polygon, or its dimensions!

Thanks to Chris and Raees for pointing out the poor name for the record.