Fixed false positive test cases for Minefield.GetHint
When changing the coordinate system to use an origin in the top left, and for axes to be 0-based I had overlooked the `GetHint` method was constructing cells with `Row` and `Column` passed in the wrong order. However, the unit tests were _also_ passing these in the wrong order to the `GetHint` method so the tests were still green!