Made implementation of GetHint clearer
We can generate the list of cells adjacent to the cell we're checking for. That makes it easier to filter off cells that don't appear in our minefield at that point, rather than when we're counting how many of those are mined.

By introducing overloads of e.g. `HasCell`, `IsMined`, and `GetHint` that accept a `Cell` rather than a row and column we can further shorten the lambda expressions as we don't have to construct a `Cell` in them now.