Added some very basic parameter validation to the IsMined method
This now allows us to see why it was important for the `GetHint` method to filter off cells that don't exist in the minefield in some fashion. If it does not do so it now causes `ArgumentException`s because it is leaning on the `IsMined` method (which may now throw).

This also exposed a bug in the `HasCell` method. We'd checked for indexing off the right or bottom of the board, but not the top or left!