Implemented behaviour to initialise a Game based on a Minefield we have already set up
The `Game` class is intended to track how the individual has played i.e. which cells have they tried to flag, have they cleared certain cells, or won or lost the game.

The `CellState` type is (gasp) intended to represent the state of any given cell. We've chosen to use the record type (rather than class or struct) so we get all the niceties around comparison and equality. This makes it easier for us to write robust tests e.g. to ask, "Is this cell in this state?"

See notes in the docs folder for more details about how I'm planning on getting through to the end of this.