Removed ambiguity around terms "Row" and "Column"
In some cases I was using this to mean how many rows or columns there are. In other cases I was using to mean an index for the row or column on the grid. Clarified this by being more explicit and using e.g. `NumberOfRows` and `RowIndex` as appropriate.

This will (might?) become more important later on when `Row` would _also_ be used to refer to the list of `Cell`s that make up that row!