Modified benchmarks to be a fairer comparison of which algorithm for GetHints is the better one
I had failed to account for having hard-wired values for the numbers of rows and columns on our fake minefield. The result was the combination of `GetAdjacentCells` and `HasCell` would filter off cells that were actually on the minefield.

I don't think this have affected the outcome we were testing for a mine whose adjacent cells existed entirely with the hard-wired values we were using. However, considering we know the size of the fake minefield we created (by virtue of the number of mines we were asked to place) we might as well make use of that data!