Wrote benchmarks to explore four ways for generating an enumerable of cells
Most of the time we only have the number of columns and the number of rows. We need to turn this into an enumerable of cells. There are four options I could think of:

* A nested `for` loop
* A nested `foreach` loop
* A `SelectMany` call
* A query expression

Although the nested `for` loop has a bit more code going on it proved to be the fastest in all four scenarios. I would assume a significant proportion of that is being the only option that _doesn't_ need to generate two `IEnumerable<int>`s.

The end result of all that is not needing to change the loop I had in place! 😂