Experimented with features of BenchmarkDotNet for combinatorial benchmarks
The idea was to see how the `IsWon` method behaves when we change the size of the minefield, the number of mines, and whether the game is won, lost, or in progress. This seemed like a good excuse to use `ParamsAttribute` and `ParamsAllValuesAttribute` to set up some combinatorial benchmarks.

Because we need to set up different `Minefield`s and "games" this seemed a good excuse to also explore `GlobalSetUpAttribute`; we don't want creating the "game" to be part of the benchmark.

For no we have replicated the behaviour that is currently in the `Game` class.

This shows I might need to re-think how I'm arranging my benchmarks into classes.