Used default interface methods to make IMinefield inherit IEnumerable<Cell>
This makes the code in `Game` and `GameWonBenchmark` a little nicer because the don't have to have the nested `for` loop. We can restrict that to `IMinefield`.

Unfortunately this makes `GameTests` a little nastier. They now have to fake the enumerator because the Fake It Easy fakes don't pick up default interface methods. These tests might now be better off using a real `Minefield` instead of the faked version.

As I'm thinking of restructuring how I hold data about where mines are, what state cells are in etc I'm going to leave it for now...