Started experimenting with treating e.g. "A1" as a "Location"
Although we _could_ represent the location as a string we have a proper type; we don't want to fall into the "primitive obsession" trap!

The idea was to remove some of the awkwardness around constantly thinking things like, "Is this index a row or a column?", and "Which order am I supposed to pass these again?".

I could try passing the location in the same format we would use on something like a chess board.

Having a new type would then also free up the name `Cell` which I was starting to think needed to be much more than just the pairing of column and row index.

Some possible downsides of this approach... We'll almost certainly end up putting more strain on the heap because we've replaced one `int` with a `string`. We've also removed the possibility of negative coordinates. Although we could have e.g. "-A-1" (which looks awful IMO) we'd still have no reasonable way to represent the 0 index for column!