Added benchmarks for possible implementations of parts of Location
We could test the validity of the `column` parameter by either enumerating the `string` as `char[]`, or by testing against a `Regex`. The former allocates a few bytes, but is _much_ faster. This is particularly so on the short strings we would expect to be handling in most cases.

For implementing `TryParse` we could catch any potential exception thrown by `Parse` but that rather defeats the purpose of the pattern! By testing against the `Regex` we have defined, then invoking the ctor we get _much_ faster performance with less memory overhead for cases when the input cannot be parsed. They are similar when the input can be parsed.