Implemented ability to convert between int and string representations of columns
We have _at least_ two scenarios where we need to convert between the two representations - when comparing two columns for adjacency, and when building an `IEnumerable<Location>` based on a number of columns and rows.

At the moment this conversion logic is being split between multiple types. This is starting to bring it into the same place.

At the moment we only deal with a single character and have no exception handling in place.