Renamed ColumnConverter to ColumnNameConverter
This more accurately conveys it is the _name_ of the column we're converting, not the actual column! Although we don't have a type to represent the column yet we very well may do. Even if we don't the name was still misleading.

Includes renaming some test methods to correctly refer to column name, rather than column string.