Implemented TryParse methods for ColumnName
Explict casts from ints or strings _might_ throw, so this gives the caller the option of _not_ taking the hit of an exception being thrown.

This means we have tests around e.g. "Is this string valid" a _third_ time. They now live in the explicit conversion operator, the `TryParse` method, and the ctor. In the same fashion we have duplicated the logic that ints have to be positive, non-zero.