Used terms referring to column and row consistently in Location
In some places I was using "Column Name", and "Column". I was also using "Row".

It seemed "Column Name" was more appropriate; it was not an actual column, but how we were referring to it.

If we started referring to a "Column Name", then it made sense to refer to a "Row Index". In much the same way it was not an actual row, but how we were referring to it.