Changed Deconstruct to output into ColumnName and int, rather than string and int
It would have been nice to have both, but we cannot overload `Deconstruct` based on parameter type; it has to be on the number of parameters. In this case it didn't make sense to have different numbers of parameters.

We can implicitly convert from `ColumnName` to string, but going the other way requires explicit conversion. The consumer also has a bit more behaviour available if dealing with a `ColumnName` rather than a string. These two combined made deconstructing into `ColumnName` seem more sensible.

Includes adding test(s) for the `Deconstruct` method.