Crude switch of Minefield to use Location rather than Cell
This leaves a lot of things that consume the IMinefield interface not compiling. The "crude" part refers to either not building projects (ConsoleUI and Benchmarks) or excluding problematic files if we _have_ to build certain projects.

To make life easier we have dumped some overloads from the `Minefield` type that were not on the interface, but might eventually put back. For example we used to be able to get a hint by passing row and column index, rather than a `Cell`. We might want to put an overload where you can pass the location string.