Added test case for GetHint a mined Location
I eventually decided why should a mined `Location` have a hint any different to an unmined one? The hint is just the count of mines in adjacent `Locations`. I can't think of a good reason why the presence of absence of a mine in the test `Location` should change that.