Modified MinefieldBenchmark to be more reflective of Minefield now using Location instead of Cell
One of the benchmarks is no longer fully reflective of what Minefield is actually now doing. It has been left because it is still a valid approach `Minefield` _could_ use, rather than leaning on `Location` and `ColumnName`.

This required putting the Benchmarks project back into the solution. However, there are still some files with compilation errors. Because we don't want to have to fix those to be able to check what we've done with `Minefield` benchmarks.