Update Minefield to use Location instead of Cell (#1)
* Crude switch of Minefield to use Location rather than Cell

This leaves a lot of things that consume the IMinefield interface not compiling. The "crude" part refers to either not building projects (ConsoleUI and Benchmarks) or excluding problematic files if we _have_ to build certain projects.

To make life easier we have dumped some overloads from the `Minefield` type that were not on the interface, but might eventually put back. For example we used to be able to get a hint by passing row and column index, rather than a `Cell`. We might want to put an overload where you can pass the location string.

* Corrected test name

When swapping `Minefield` to use `Location` instead of `Cell` I had miss at least one reference to `Cell` 😢

* Added test case for GetHint a mined Location

I eventually decided why should a mined `Location` have a hint any different to an unmined one? The hint is just the count of mines in adjacent `Locations`. I can't think of a good reason why the presence of absence of a mine in the test `Location` should change that.

* Fixed issue where Minefield.IsMined could incorrectly throw ArgumentException

The method would sometimes throw an `ArgumentException` stating the passed `Location` was not in the `Minefield` even when it was.

This was caused by switching from a 0-based coordinate system to a 1-based one, but not updating the logic to account for this.

* Updated MinefieldGenerator to account for Minefield now using Location instead of Cell

* Changed MinefieldGenerator to return IMinefield, rather than Minefield

* Updated MinefieldGenerator to account for Minefield now using Location instead of Cell

* Updated comment explaining how we need to make tests for MinefieldRenderer more focussed

* Added ConsoleUI back into the build now MinefieldGenerator and MinefieldRenderer have been fixed

* Updated Game to account for Minefield now using Location instead of Cell

Game itself required barely any changes 😀 More changes were required in the test fixture, but they were all related to the setup of the fake `IMinefield`. This adds further weight to my thought process we'd be better of turning these into "module" tests and just using a real `Minefield`.

* Modified MinefieldBenchmark to be more reflective of Minefield now using Location instead of Cell

One of the benchmarks is no longer fully reflective of what Minefield is actually now doing. It has been left because it is still a valid approach `Minefield` _could_ use, rather than leaning on `Location` and `ColumnName`.

This required putting the Benchmarks project back into the solution. However, there are still some files with compilation errors. Because we don't want to have to fix those to be able to check what we've done with `Minefield` benchmarks.

* Improved accuracy of benchmarks for Minefield

At the time I wrote this I didn't know about how to do set-up for benchmarks. Code that should be set-up and _not_ benchmarked (i.e. the creation of our "minefield") has now been moved into `GlobalSetup`.

* Fixed default implementation of GetLocations on IMinefield

This had not been updated to account for the old coordinate system being 0-based on both axes to the new coordinate system being 1-based. As such starting at (0,0) was guaranteed to throw an exception! 🤦‍♂️

* Updated benchmarks for IsWon to account for us of Location instead of Cell

`Game` utilises `Minefield` for knowledge of whether certain locations are mined. `Minefield` now utilises `Location` instead of `Cell`. The benchmarks have been updated to account for that.

* Updated Game benchmarks to account for use of Location instead of Cell

These benchmarks are slightly out of date insofar as they are dealing with behaviour that is now part of the default implementation of `IEnumerable<Location>` that exists on `IMinefield`. They are, however, still valid benchmarks, and will be moved across in due course.