Added a Contains method to IMinefield, and a default implementation
A potential problem with this approach is if people use a type that only implements `IMinefield` they pick up the extension method `Contains<T>` in the `System.Linq` namespace. The consumer has to type their variable as `IMinefield` to get the default implementation. This is IMO non-obvious.

The obvious solution is to rename the `Contains` method to something like `ContainsLocation`, but that feels a bit clunky 😢