Added benchmarks for the default implementation of Contains on the IMinefield interface
Although a little awkward to get at (for now) these benchmarks prove it is definitely worth doing so. Our implementation _massively_ outperforms the LINQ one in _all_ circumstances. Unlike the LINQ version its performance is also linear. The LINQ version is slower based on the passed location, _and_ slows down dramatically as the size of the minefield increases.

I suspect the abysmal performance of the LINQ implementation is due to rebuilding the enumerable on each call. I'm confident it would perform better if e.g. Minefield stashed a List<Location> and returned that each time. However, I'm also confident it would _still_ be slower than our version that implements the test with some arithmetic.