Added Size property with default interface implementation to IMinefield
This includes benchmarks which demonstrate this is (as you would expect) _massively_ faster than using the `Count` extension method. This is because we never have to touch the `IEnumerable` interface. Not only are we faster, we have less allocations and garbage collection, as well as performance being the same regardless of the size.

A slight annoyance is even though we use categories to restrict which benchmarks we run the presence of `Position` means we end up running each benchmark three times. That's useful for `Contains`, but serves no benefit for `Size`.