Repurposed the Cell type
Initially `Cell` was to be a representation of the coordinates, and we would have a state machine implementation. There would be a dictionary of `Cell`s to `CellState`s. This had presented a number of irritations such as not having a good way to return a hint when revealing something, and slow calculations of whether a game was won or lost.

We had already replaced the use of `Cell` with `Location`. That meant we could repurpose Cell to hold information such as whether that location had been revealed and (in future) what the hint for that `Cell` would be.