Improved performance of Game.IsWon and Game.IsLost
The switch of `Game` to store a dictionary of the new `Cell` rather than `CellState` meant deriving `IsWon` and `IsLost` were performing significantly worse than they had previously (even though we were getting a number of other benefits).

By switching over to storing and maintaining state information, rather than deriving it, we get much faster performance, and the execution time remains static regardless of how many cells, mines, or whether the game is won or lost. In addition we now execute in about 1% of the time we used to before the switch to `Cell` (as you'd probably expect).

We've had to set our data up in a slightly clunky way to start with. This is another consequence of a FakeItEasy fake not picking up default interface implementations.