Changed tests for Game to use a real Minefield
Although no longer what I would think of as unit tests this was (IMO) the pragmatic choice. The tests for `Game` were quite clumsy due to the amount of set-up work we were having to do. In addition, we'd had to accept a sub-optimal implementation of `Game` in order to avoid having to do _even more_ set-up. The tests are arguably more intention-revealing now they no longer have a lot of noise in them.