Exposed Hint on Cell and ensured Game initialises it correctly
Exposing this data on the `Cell` type means clients will now have an easy way to render this information when a `Cell` is returned by calling `Reveal`.

Deriving this once during construction ensures we'll get O(1) performance from that point on; we'll take the hit on construction for that benefit.