Changed Minefield.GetHint to find adjacent Locations then count mined cells
The previous implementation found every mined `Location` then checked to see if it was adjacent to the test `Location`. Our benchmarking showed this was faster for games with up to 7 mines. Once we got to 8 mines or more determining the adjacent 'Locations' then checking to see if they are mined is faster. It is also a O(1) function, whereas the previous implementation was 0(n).

We can probably make the determination of the adjacent locations faster by avoiding the LINQ expressions.