Made Minefield.GetHint faster by selecting appropriate strategy based on number of mines
We had seen from benchmarking getting the hint for a `Cell` was faster when there were few mines if you counted the mined cells that were adjacent. Once you got up to 8 or more mines it became faster to find the adjacent cells then see which had a mine in them.