Added parameter validation to GameRenderer.RenderRow
I'm not sure if `InvalidOperationException` is the right thing to do when `rowIndex` is greater than `numberOfRows`. The exception is related to the arguments, but it is the arguments _in combination_ rather than a specific argument.

It is also becoming apparent that I should have been using `unit`, rather than int, considering the number of places I'm validating for not negative.