Renamed GameRenderer to NaiveRowRenderer and extracted interface
This name more accurately reflects the type renders a row and various supporting things such as borders; it does not actually render the `Game`. We have used the prefix `Naive` on the name because we've chosen the naive implementation with strings and made no attempt to optimise performance or memory consumption.

This frees up the name `GameRenderer` for a type that will effectively coordinate calls to an `IRowRenderer`