Renamed the Benchmarks project to Engine.Benchmarks
This is because we will shortly be getting a series of benchmarks for types in ConsoleUI. We wouldn't want to intermingle these ones with benchmarks for the engine itself.

This includes correcting the namespace to match that of the types we are benchmarking. This is consistent with how we set the namespaces for the test assemblies.