Tried implementation of rendering top border using string.Create
This implementation is _significantly_ faster than the naive one, especially as the grid gets bigger. It is also more memory efficient.

This is a nice improvement even before we account for an implementation of `IGameRenderer` would be able to cache the response and re-use it; it will never change no matter how many times we render the game.

We may find further improvements with exactly how we're filling the `Span<char>`, and how we're getting the length of the prefix.