Moved implementation of RenderTopRow using string.Create into a new implementation of IRowRenderer
The type is actually an `IGameRenderer`, but we want to force an implementation of `IRowRenderer` so we know we get an efficient one.

Benchmarks were renamed because we're now comparing implementations of `IRowRenderer` to each other, as well as other implementations that aren't housed in a specific type.

This is already showing we're going to end up with two test fixtures that have identical tests.