Changed signature of IRowRenderer.RenderRow to expect the ICellRenderer to be passed
This means we can keep the constructor of NaiveRowRenderer as parameterless. In turn that is intended to make removing duplication between tests for `NaiveRowRenderer` and `EfficientGameRenderer` a lot easier.

This is also much more in keeping with the sort of signature you would actually expect; the only method on the interface that requires the `ICellRenderer` is `RenderRow` so there is no need to make the entire object dependent upon it!