Removed duplication from tests for two implementations of IRowRenderer
We want to ensure the new implementation that manages strings more efficiently (`EfficientGameRenderer`) has the same outputs as `NaiveRowRenderer`. However, we didn't want to duplicate every test. We could achieve this by having tests accept the instance of `IRowRenderer` to test against, rather than creating their own.

This technique was reliant on any dependencies being passed to methods rather than constructors. That was to allow the test method to set up any behaviours on fakes.