Tried alternative implementations of IRowRenderer.RenderTopBorder
Manually creating a `char[]` and populating it has proved faster than using `string.Create`, albeit at the cost of more memory and garbage collection as the size of the minefield used by the `Game` increases.

Unsurprisingly, using an `IEnumerable<string>` and `string.Join` performs abysmally both in terms of speed and space.