Removed pre-filling Span<char> with a common character
Replaced this with only setting the exact character inside of the loop. This cuts out a number of superfluous assignments where we would just be over-writing the character later.

This has, as expected given us a small improvement in speed, but space remains the same. This is still slower than using a `char[]` until we get up to a _very_ large number of rows and columns; [Minesweeper Online](https://minesweeperonline.com/) defines the Expert level as 16x30 and the `char[]` implementation is still faster at 100 mines.