Implemented RenderRow in EfficientGameRenderer
As with rendering the box art we have used `string.Create` and a `SpanAction` to avoid excessive allocations, and LINQ enumerations in the form of a `Select` method.

The code is not particularly clear when we're having to do plenty of arithmetic with loop variables, and array lengths to calculate an indexer, but I guess that is the sacrifice we make for the performance gains.