Added benchmarks for the two implementations of RenderRow
Somewhat surprisingly this showed the (supposedly) efficient version was not actually better than the (supposedly) naive version. Presumably, this is because the bulk of the time is spent managing the cells, rather than creating the string. In order to avoid multiple enumerations the `EfficientGameRenderer` makes a copy of the `IEnumerable`. My suspicion is this offsets any advantage we get from using `string.Create`.