Improved accuracy of benchmarks when testing RenderRow
The benchmarks had a major problem in the creation of the `IEnumerable<Cell>` being included in the benchmark. The time required for this vastly outweighed the time to actually render the row. That meant all implementations _appeared_ to be very similar in terms of speed and space!

Moving the creation of the `IEnumerable<Cell>` to `GlobalSetUp` means we can see the `EfficientGameRenderer` implementation _is_ much more efficient, typically executing in only about 20 to 25% of the time, and using less than half the memory.

The changes to the benchmarks also show that creating the `char[]` directly is even more efficient than using `string.Create`. This is consistent with what we saw with `RenderTopBorder`.