Improved memory efficiency of EfficientGameRenderer.RenderRow
We could avoid a few heap allocations by not manually copying the `IEnumerable<Cell>` into a `List<Cell>` which we were doing to avoid multiple enumerations, and so we could index into variable.

By rendering the `Cell`s outside of the loop we could simplify the state tuple we were passing, avoiding the need to pass the `ICellRenderer`.

There was also no need to manually turn the row index into a `char[]` as we can index into a `string`.