Corrected some "cheating" for benchmarking rendering a row using a char[]
The inline implementation had been using a `List<char>` directly. _If_ this were ever in a real implementation that would not be possible; it would only have access to an `IEnumerable<Cell>`. Modified the benchmark to account for a real implementation would have to copy that into something it could index in some form.

These changes have suggested using a `char[]` directly would not be much fast than `string.Create`, but would typically consume more memory, especially when the number of rows starts to go up.