Added benchmarks for the two implementations of RenderColumnNames
In the worst case scenario the efficient version executes in about 50% or the time, and only consumes 2/3 of the memory. In the best case it executes in about a quarter of the time, and only consumes half the memory. The more columns there are better the efficient implementation performs.

This is obviously a significant improvement, but not to the extent we saw with `RenderTopBorder` and `RenderRow`. Presumably this is because there is a lot of overhead involved in getting the `ColumnName` represented as a `string` and both implementations will have to take that hit.