Implemented CharArrayGameRenderer
This is an implementation of `IRowRenderer` using a `char[]` directly, rather than `string.Create`. Benchmarks have shown this executes a tiny bit slower than `StringCreateGameRenderer`, _and_ consumes more memory. That is a bit of a surprise as the behaviour is a little different to what we saw before putting this code into its own class.

This is still worth keeping. Because we now have the code for getting the `char[]` for each line in the full game output we can make a very fast implementation of `IGameRenderer` by keeping hold of `char[]` and only replacing the actual cells, rather then re-rendering the entire grid!!

Much like with `StringCreateGameRenderer` although this implements `IRowRenderer` we have named it as a game renderer because that is primarily how we intend to use it.