Added benchmarks for implementations of IGameRenderer
This has show that for now we don't see much difference in performance between the two 😢 This is in spite of us knowing the `IRowRenderer` implementation in `StringCreateRenderer` is much faster than `NaiveRowRenderer`; we've benchmarked them. We also know we have far fewer operations because we only render the row separator once for `StringCreateRenderer`, rather than for every row in `NaiveGameRenderer`.

The lack of difference in speed has been tracked down to the bulk of the execution time coming from enumerating the `ColumnName`s and `Cell`s via the passed instance of `Game`.

On the plus side `StringCreateGameRenderer` is consuming a _tiny_ bit less memory! 😂

Approximately 14 seconds to render a 1,000 x 1,000 grid is not acceptable. Just over 1/10 of a second to render a 100 x 100 grid just about it.